Rework after review
