+++ model/infinispan/src/main/java/org/keycloak/cluster/infinispan/DatabaseAwareClusterProvider.java
@@ -26,9 +26,9 @@ import java.util.concurrent.Future;
 import org.keycloak.cluster.ClusterEvent;
 import org.keycloak.cluster.ClusterEventStoreProvider;
 import org.keycloak.cluster.ClusterListener;
-import org.keycloak.cluster.ClusterLockStoreProvider;
 import org.keycloak.cluster.ClusterProvider;
 import org.keycloak.cluster.ExecutionResult;
+import org.keycloak.cluster.jpa.ClusterLockStoreProvider;
 import org.keycloak.connections.infinispan.NodeInfo;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.KeycloakSessionFactory;
