+++ model/jpa/src/main/java/org/keycloak/cluster/jpa/ClusterLockStoreProviderFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Red Hat, Inc. and/or its affiliates
+ * Copyright 2026 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,33 +13,12 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package org.keycloak.cluster;
+package org.keycloak.cluster.jpa;
 
-import org.keycloak.provider.Provider;
 import org.keycloak.provider.ProviderFactory;
-import org.keycloak.provider.Spi;
 
-public class ClusterLockStoreSpi implements Spi {
-
-    @Override
-    public boolean isInternal() {
-        return true;
-    }
-
-    @Override
-    public String getName() {
-        return "cluster-lock-store";
-    }
-
-    @Override
-    public Class<? extends Provider> getProviderClass() {
-        return ClusterLockStoreProvider.class;
-    }
-
-    @Override
-    public Class<? extends ProviderFactory<?>> getProviderFactoryClass() {
-        return ClusterLockStoreProviderFactory.class;
-    }
+public interface ClusterLockStoreProviderFactory extends ProviderFactory<ClusterLockStoreProvider> {
 }
