+++ model/jpa/src/main/java/org/keycloak/cluster/jpa/JpaClusterLockStoreProvider.java
@@ -22,7 +22,6 @@ import java.util.List;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.LockModeType;
 
-import org.keycloak.cluster.ClusterLockStoreProvider;
 import org.keycloak.common.util.Time;
 import org.keycloak.connections.jpa.JpaConnectionProvider;
 import org.keycloak.models.KeycloakSession;
