+++ model/jpa/src/main/resources/META-INF/services/org.keycloak.cluster.jpa.ClusterLockStoreProviderFactory
