+++ model/jpa/src/main/resources/META-INF/services/org.keycloak.provider.Spi
@@ -19,3 +19,4 @@ org.keycloak.connections.jpa.JpaConnectionSpi
 org.keycloak.connections.jpa.updater.JpaUpdaterSpi
 org.keycloak.connections.jpa.updater.liquibase.conn.LiquibaseConnectionSpi
 org.keycloak.connections.jpa.entityprovider.JpaEntitySpi
+org.keycloak.cluster.jpa.ClusterLockStoreSpi
