+++ server-spi-private/src/main/resources/META-INF/services/org.keycloak.provider.Spi
@@ -35,7 +35,6 @@ org.keycloak.migration.MigrationSpi
 org.keycloak.events.EventListenerSpi
 org.keycloak.events.EventStoreSpi
 org.keycloak.cluster.ClusterEventStoreSpi
-org.keycloak.cluster.ClusterLockStoreSpi
 org.keycloak.exportimport.ExportSpi
 org.keycloak.exportimport.ImportSpi
 org.keycloak.timer.TimerSpi
